proposal for a much shorter abstract and intro
I actually also think we can remove the other subsections in the intro. However, we could also check if we want to keep some of it or just move it to the appendix but there is not a huge amount of value because all the text can still be found in the RFC for scream v1.