Add high-level equations to overview
Proposing this as I thought it might help understand the algorithm on a high level.

However, I'm not fully sure about the decrease part. I added qtarget_alpha, which is not defined yet, because it looks like there is only multiplicative decrease but I have to say the whole target_delay related part is a bit confusing to me right now. Need to double-check!

Further, it could be good to give some more hinds about the increase part as well but not sure how to summarise it easily in one line of pseudo code...

Do it help to add this?